refactor(dfpt): split DFPT_Phon Ewald and electronic terms into own files
- move ion_ion into dfpt_phon_ewald.cpp with the alpha search, the
  reciprocal-space sq/s0 + pair sums, the shared erfc h_ab tensor and
  the real-space self-image/pair sums as file-local helpers
  (cyclomatic complexity 38 -> main body 4)
- move accumulate_electron into dfpt_phon_elec.cpp with the dpsi
  stash/restore, the bare-potential cross sum and the chi scatter as
  helpers; the same-atom d2 block becomes the private member
  accum_d2_same_atom (complexity 34 -> 5)
- all formulas and empirical parameters moved verbatim; wire the two
  new translation units into the 3 CMakeLists

Verification: make -j 30 clean (build and build_max_para_test);
ctest MODULE_DFPT_phon_serial/q0_serial/pw_run/pw_data 4/4 passed;
code_quality_score.py: dfpt_phon.cpp 34 -> 100, new files 100,
dfpt_phon.h 98.
