+++ source/source_pw/module_dfpt/test/CMakeLists.txt
@@ -44,6 +44,8 @@ AddTest(
 	../dfpt_stern.cpp
 	../dfpt_rho.cpp
 	../dfpt_phon.cpp
+	../dfpt_phon_ewald.cpp
+	../dfpt_phon_elec.cpp
 	../dfpt_q0.cpp
 	../dfpt_q0_pos.cpp
 	../dfpt_metal.cpp
